+++ getTraces_single_connection_loop.py
@@ -26,7 +26,8 @@ import numpy as np
 
 # Default options
 VISA_ADDRESS = 'USB0::2391::6038::MY57233636::INSTR' # address of instrument
-WAVEFORM_FORMAT = 'BYTE'    # BYTE formatted data is transferred as 8-bit bytes.
+WAVEFORM_FORMAT = 'WORD'    # WORD formatted data is transferred as 16-bit uint.
+                            # BYTE formatted data is transferred as 8-bit uint.
                             # ASCii formatted data converts the internal integer data values to real Y-axis values.
                             #       Values are transferred as ASCii digits in floating point notation, separated by commas.
 CH_NUMS=['']        # list of chars, e.g. ['1', '3']. Use a list with an empty string [''] to capture all currently displayed channels
